Create SUPPVIEW.COB
View module of the Suppliers record. At the moment it has 2 options which are show all records or one specific. All records shows one by one after each enter press. The other asks for the supplier ID and shows the record if it exists. All tested and working but on the to-do list still need other ways of finding specific records that are easier for the user. #24